requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.1.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.7.0
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-environ==0.9.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==9.0.3
progressbar2==4.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.4.2
stripe==12.2.0
typing_extensions==4.13.2
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.6.0
